Made new commands for automatic operation, changed certain payment schemes, fixed holding contract being unable to run at epoch 0

+++ src/main/scala/persistence/responses/ShareResponse.scala
@@ -4,5 +4,5 @@ import org.ergoplatform.appkit.Address
 
 import java.sql.Date
 
-case class ShareResponse(poolId: String, height: Int, diff: Double, netDiff: Double, minerAddress: String, source:String, created:Date)
+case class ShareResponse(poolId: String, height: Long, diff: BigDecimal, netDiff: BigDecimal, minerAddress: String, source:String, created:Date)
